VIS-1023: Fix newline in node tooltips and add test (#2)
* Fix issue with spacing in hover test

* Fix newline in node tooltips and add test

* Specify node version

* Add test to travis

* Fix test
