+++ .travis.yml
@@ -1,2 +1,9 @@
-language: r
-cache: packages
+language: node_js
+node_js:
+  - '11.13'
+
+install:
+  - npm install
+
+script:
+  - npm run test
