+++ package.json
@@ -1,0 +1,14 @@
+{
+  "name": "networkD3",
+  "private": true,
+  "engines": {
+    "node": ">=12.9.0"
+  },
+  "scripts": {
+    "test": "jest"
+  },
+  "dependencies": {
+    "jest": "^27.3.1",
+    "puppeteer": "^11.0.0"
+  }
+}
